meta: Add model based nodes
It is currently only working for the sources, not the sinks.

The general idea is to convert the object node into a model node
by adding a "QPropertyToModel" model to this projet. It could also
support signals, slots and hybrids using concatenation model proxies.

This is superior to the current QObject node implementation because:

1) it is simpler (as long as you understand models
2) it is more flexible (you can extend it using proxies
3) it is easier for the node to track its connections
4) it removes duplicated logic from the code

It could, once completed, become the "base" node implementations
and cut the project size in half.

Note that this code is a working snapshot of a work in progress
feature. It is good enough as is to be used, but the code is hacky.