edge: Turn the QGraphicsItem into a property rather than base class
This is the second step in the planned MVC conversion. The public
API and the node connection internals wont be mixed up anymore.

This will allow to turn the connections themselves into QModelIndex
pairs and the sockets into a QAbstractItemModel proxy.

This commit also remove unused constructors. Most of them made no
sense anyway.