node: Move the GraphicsItem from base to method
This makes the API a little awkward to use, as `addItem` is still the
official way to add to the scene. This will fixed once the scene and view
are moved to the "lets make a nice library API" conventions.

This will however wait as my priority is the model part (I need it).