model: Begin to create a proper model layer
As mentionned in the previous commits, the plan is to turn this node
editor into a compliant Qt Model/View system. Instead of creating our
own little data layer, this will extend Qt one.

This commit add the first (of many) components required to do that. It
adds a model to convert any QObject into a QAbstractItemModel.

Its properties becomes QModelIndex. Setters and getter are using
data() and setData(), notify use dataChanged and so on. It isn't
fully implemented yet, but good enough to move on with static tests.

The next step is to add a new "manager" model that concatenate all
nodes individual models into a tree model where the top level indices
are the nodes and the children a proxy to the node own model.

After that, a new proxy model has to be developped to connect different
QModelIndex dataChanged to the other setData. It also has to generate
a "connections model" storing the links.

Then, an internal extended QIdentityProxyModel add the "edge"
GraphicsItem concept.

Finally, turn the GraphicsNodeView into a QAbstractItemView, hide the
scene and rename everything. A perfect Qt node edition framework!