model: Add the QMultiModelTree model.
Takes the first column of multiple source models and create a tree
for them.

This is currently as basic as it gets, it doesn't track source changes
at all. Keep in mind this is a step in the "master plan", it doesn't
have to do anything else at this point and it would be a pain to test
anyway.

This is the backend of the node editor data layer. On top of this
will be a QIdentityProxyModel based class to handle the connections
between the QModelIndex (node sockets).

After this will be refactored into the node editor core, this model
will have to be completed, along with the QObjectModel.