data layer: Remove deprecated connection methods
More will eventually be removed, but this commits forces more
code paths to go through he data layer instead of doing thier own
thing.

It is a work in progress feature. But this commit brings back some
of the features lost in the inversion of responsability (edges are
working again).