socketproxy: Force the QObjectModel value role to be synched.
This is a hack to "fix" the basic use case. It is now possible
to play with QObject based nodes again.

There is tons of unhandled corner cases, but it "works" again in
the sense it's possible to add nodes and connect their property:

 * Only either as source or sinks
 * Disconnection doesn't work
 * dataChanged only work if a single element changed
 * edges are still not fully using the data layer
 * sockets can only be appended, not removed
 * no other use case than some simple qobject works
 * memory is leaked

Still, this is another big milestone and prove the current design
can work. It is only a matter of fixing all bugs and corner cases.