data layer: Complete the edge transition
Edges are now fully handled by the data layer.

This commit also move the position managing to the data layer.
This was necessary to uncouple the view and socket from the edge.

More can now be moved to the data layer, such as the node, socket
and edge title. There is still many edge cases not handled, such
as deleting disabled connections (and their edges).

More code is now "partially dead" and can be removed, but this can
be done in its own commit.