edge: Fix disconnection and simplify the process
Now that connections are handled by the data layer, the view doesn't
need to be "smart" anymore.

The GraphicsView part has been dumbed down to preventing it from
breaking the disconnections.

Note: The data layer is now on par with the old code. It has the same
bugs and limitations the old code had, but at least there is a
clean path to address them (see all the TODO).