socketModel: Make progress on finxing the missing data layer links
It should be possible to:

 * go from and edge to both nodes
 * go from a socket to its edge
 * go from a node to its edges
 * do from a socket to its sockets
 * go from a socket to its node and sibling node (if any)

It should also be possible to keep track of the edge dependencies when
the model is dynamic (adds and removes sockets during execution)

This commit wrap the sockets in an internal tracking structure and
adds a proxy model to get the sinks or sources for a node.

This is a work in progress, the proxy model still needs to track the
connections. However as this is a stable point in the development,
better commit now.